move comm listener into its own module
