+++ extensions/positron-r/amalthea/crates/amalthea/src/comm/mod.rs
@@ -6,4 +6,5 @@
  */
 
 pub mod comm_channel;
+pub mod comm_listener;
 pub mod lsp_comm;
